requirements.txt
absl-py
seaborn
pandas
toolz
imageio
